backend/requirements.txt
# Do not use pip freeze for Docker environments 
# - always manually maintain your requirements.txt with verified versions.
asgiref==3.8.1
colorama==0.4.6
Django==4.2.12
django-environ==0.12.0  
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
sqlparse==0.4.4
tzdata==2024.1
django-storages==1.13.2
boto3==1.34.94
django-cors-headers>=4.0
PyJWT>=2.0.0
celery==5.3.6
redis==5.0.1
celery[redis]==5.3.6
django-environ
# For agents system
pdfplumber
langchain
pinecone-client
google-cloud-aiplatform
transformers==4.38.1
torch==2.2.2
httpx>=0.25.0
